refactor(circuit): Remove circuit list widget to correct entity widget folder
- Made circuit entity's named widget folder
- Removed CircuitList to circuit widget ui segment
- Fixed CircuitList public api
