+++ packages/client/src/app/circuits/page.tsx
@@ -5,7 +5,7 @@ import { getBreadcrumbsItemsCircuits } from '~app/circuits/breadcrumbs';
 
 import { useCircuitsServer } from '~entities/circuit/api';
 
-import { CircuitList } from '~widgets/circuit-list/ui/CircuitList';
+import { CircuitList } from '~widgets/circuit/ui';
 
 interface ICircuitsPage {
     searchParams: {
