+++ packages/client/src/widgets/circuit/ui/index.ts
@@ -1,0 +1,1 @@
+export { CircuitList } from './list';
