+++ packages/client/src/widgets/circuit/ui/list/index.ts
@@ -1,0 +1,1 @@
+export { CircuitList } from './CircuitList';
